# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GitHub Bounty API",
    "description": "API for searching and discovering GitHub bounty programs and issues",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/search/issues": {
      "get": {
        "summary": "Search Bounty Issues",
        "description": "Search GitHub issues that might have bounties\n\nThis endpoint searches GitHub issues with common bounty labels and keywords.",
        "operationId": "search_bounty_issues_api_search_issues_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search query string",
              "title": "Query"
            },
            "description": "Search query string"
          },
          {
            "name": "labels",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of labels",
              "title": "Labels"
            },
            "description": "Comma-separated list of labels"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Programming language filter",
              "title": "Language"
            },
            "description": "Programming language filter"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Issue state (open/closed)",
              "default": "open",
              "title": "State"
            },
            "description": "Issue state (open/closed)"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Results per page",
              "default": 30,
              "title": "Per Page"
            },
            "description": "Results per page"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BountyIssue": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "repo_name": {
            "type": "string",
            "title": "Repo Name"
          },
          "repo_url": {
            "type": "string",
            "title": "Repo Url"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "comments_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments Count"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url",
          "repo_name",
          "repo_url",
          "labels",
          "state",
          "created_at"
        ],
        "title": "BountyIssue"
      },
      "BountyProgram": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url",
          "platform"
        ],
        "title": "BountyProgram"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SearchResponse": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/BountyIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "programs": {
            "items": {
              "$ref": "#/components/schemas/BountyProgram"
            },
            "type": "array",
            "title": "Programs"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "issues",
          "programs"
        ],
        "title": "SearchResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}